Move operations back to component bounded context
Splitting operations (namely metadata and monitoring) to a
separate module outside of component's bounded context led to
overall low cohesion and high coupling.

The lifecycle of operations is the same as that of the component
(temporal cohesion). Operations needs to access and be aware of
component architecture in detail (information cohesion).