Remove AppRegistry application
Current implementation grouped resources of all environments under a single application.
I didn't find a way to filter and organize resources information per environment using that approach.
The reason I'd want to do such an organization is be able to view costs, metrics and so on per environment.
Removing this example until I find a better way to implement this.